Remove the rest of the code toggles
