+++ codeToggles.js
@@ -1,13 +1,6 @@
 // This file is a map of github issue numbers with booleans
 // TODO: move to a modifiable location such as S3 or DynamoDB
-const toggles = {
-  6936: true,
-  7142: true,
-  7178: true,
-  7184: true,
-  7198: true,
-  7199: true,
-};
+const toggles = {};
 
 exports.isCodeEnabled = issueNumber => toggles[issueNumber];
 exports.isCodeDisabled = issueNumber => !toggles[issueNumber];
