+++ shared/src/business/test/createTestApplicationContext.js
@@ -106,7 +106,6 @@ const {
 const {
   incrementCounter,
 } = require('../../persistence/dynamo/helpers/incrementCounter');
-import { isCodeEnabled } from '../../../../codeToggles';
 const {
   putWorkItemInOutbox,
 } = require('../../persistence/dynamo/workitems/putWorkItemInOutbox');
@@ -425,20 +424,6 @@ const createTestApplicationContext = ({ user } = {}) => {
     createDocketNumber: jest.fn().mockImplementation(createDocketNumber),
   };
 
-  const mockedCodeToggles = {};
-
-  const setMockedCodeToggles = (issueNumber, value) => {
-    mockedCodeToggles[issueNumber] = value;
-  };
-
-  const mockIsCodeEnabled = issueNumber => {
-    if (mockedCodeToggles && mockedCodeToggles.hasOwnProperty(issueNumber)) {
-      return mockedCodeToggles[issueNumber];
-    } else {
-      return isCodeEnabled(issueNumber);
-    }
-  };
-
   const applicationContext = {
     ...sharedAppContext,
     barNumberGenerator: {
@@ -524,7 +509,6 @@ const createTestApplicationContext = ({ user } = {}) => {
     getUseCases: appContextProxy(),
     getUtilities: mockGetUtilities,
     initHoneybadger: appContextProxy(),
-    isCodeEnabled: jest.fn().mockImplementation(mockIsCodeEnabled),
     logger: {
       debug: jest.fn(),
       error: jest.fn(),
@@ -533,7 +517,6 @@ const createTestApplicationContext = ({ user } = {}) => {
     notifyHoneybadger: jest.fn(),
     setCurrentUser: jest.fn(),
     setCurrentUserToken: jest.fn(),
-    setMockedCodeToggles,
   };
   return applicationContext;
 };
