+++ shared/src/persistence/elasticsearch/fetchPendingItems.test.js
@@ -3,7 +3,6 @@ const {
 } = require('../../business/test/createTestApplicationContext');
 const { fetchPendingItems } = require('./fetchPendingItems');
 jest.mock('./searchClient');
-const { isCodeEnabled } = require('../../../../codeToggles');
 const { search } = require('./searchClient');
 
 describe('fetchPendingItems', () => {
@@ -49,24 +48,22 @@ describe('fetchPendingItems', () => {
     const searchQuery =
       search.mock.calls[0][0].searchParameters.body.query.bool.must;
 
-    if (isCodeEnabled(7198)) {
-      expect(searchQuery[4]).toMatchObject({
-        bool: {
-          minimum_should_match: 1,
-          should: [
-            {
-              exists: {
-                field: 'servedAt',
-              },
+    expect(searchQuery[4]).toMatchObject({
+      bool: {
+        minimum_should_match: 1,
+        should: [
+          {
+            exists: {
+              field: 'servedAt',
             },
-            {
-              term: {
-                'isLegacyServed.BOOL': true,
-              },
+          },
+          {
+            term: {
+              'isLegacyServed.BOOL': true,
             },
-          ],
-        },
-      });
-    }
+          },
+        ],
+      },
+    });
   });
 });
