+++ web-client/integration-tests/journey/docketClerkEditsDocketEntryMeta.js
@@ -1,6 +1,5 @@
 import { AUTOMATIC_BLOCKED_REASONS } from '../../../shared/src/business/entities/EntityConstants';
 import { formattedCaseDetail as formattedCaseDetailComputed } from '../../src/presenter/computeds/formattedCaseDetail';
-import { isCodeEnabled } from '../../../codeToggles';
 import { runCompute } from 'cerebral/test';
 import { withAppContextDecorator } from '../../src/withAppContext';
 
@@ -57,12 +56,10 @@ export const docketClerkEditsDocketEntryMeta = (test, docketRecordIndex) => {
       value: true,
     });
 
-    if (isCodeEnabled(7178)) {
-      await test.runSequence('updateDocketEntryMetaDocumentFormValueSequence', {
-        key: 'pending',
-        value: true,
-      });
-    }
+    await test.runSequence('updateDocketEntryMetaDocumentFormValueSequence', {
+      key: 'pending',
+      value: true,
+    });
 
     await test.runSequence('submitEditDocketEntryMetaSequence', {
       docketNumber: test.docketNumber,
@@ -87,32 +84,30 @@ export const docketClerkEditsDocketEntryMeta = (test, docketRecordIndex) => {
       message: 'Docket entry changes saved.',
     });
 
-    if (isCodeEnabled(7178)) {
-      expect(test.getState('caseDetail.automaticBlocked')).toEqual(true);
-      expect(test.getState('caseDetail.automaticBlockedReason')).toEqual(
-        AUTOMATIC_BLOCKED_REASONS.pending,
-      );
-      expect(test.getState('caseDetail.hasPendingItems')).toEqual(true);
-      const docketEntries = test.getState('caseDetail.docketEntries');
-      const pendingDocketEntry = docketEntries.find(
-        d => d.index === docketRecordIndex,
-      );
+    expect(test.getState('caseDetail.automaticBlocked')).toEqual(true);
+    expect(test.getState('caseDetail.automaticBlockedReason')).toEqual(
+      AUTOMATIC_BLOCKED_REASONS.pending,
+    );
+    expect(test.getState('caseDetail.hasPendingItems')).toEqual(true);
+    const docketEntries = test.getState('caseDetail.docketEntries');
+    const pendingDocketEntry = docketEntries.find(
+      d => d.index === docketRecordIndex,
+    );
 
-      expect(pendingDocketEntry.pending).toEqual(true);
+    expect(pendingDocketEntry.pending).toEqual(true);
 
-      const caseDetailFormatted = runCompute(formattedCaseDetail, {
-        state: test.getState(),
-      });
+    const caseDetailFormatted = runCompute(formattedCaseDetail, {
+      state: test.getState(),
+    });
 
-      expect(
-        caseDetailFormatted.formattedPendingDocketEntriesOnDocketRecord,
-      ).toEqual(
-        expect.arrayContaining([
-          expect.objectContaining({
-            docketEntryId: pendingDocketEntry.docketEntryId,
-          }),
-        ]),
-      );
-    }
+    expect(
+      caseDetailFormatted.formattedPendingDocketEntriesOnDocketRecord,
+    ).toEqual(
+      expect.arrayContaining([
+        expect.objectContaining({
+          docketEntryId: pendingDocketEntry.docketEntryId,
+        }),
+      ]),
+    );
   });
 };
