+++ web-client/src/applicationContext.js
@@ -15,7 +15,6 @@ import {
   compareISODateStrings,
   compareStrings,
 } from '../../shared/src/business/utilities/sortFunctions';
-import { isCodeEnabled } from '../../codeToggles';
 const {
   filterQcItemsByAssociatedJudge,
 } = require('../../shared/src/business/utilities/filterQcItemsByAssociatedJudge');
@@ -643,7 +642,6 @@ const applicationContext = {
     };
   },
   initHoneybadger,
-  isCodeEnabled,
   notifyHoneybadger: async (message, context) => {
     const honeybadger = await initHoneybadger();
