+++ web-client/src/router.js
@@ -1,5 +1,4 @@
 import { forEach, set } from 'lodash';
-import { isCodeEnabled } from '../../codeToggles';
 import { queryStringDecoder } from './utilities/queryStringDecoder';
 import { setPageTitle } from './presenter/utilities/setPageTitle';
 import route from 'riot-route';
@@ -1136,17 +1135,15 @@ const router = {
       }
     });
 
-    if (isCodeEnabled(7142)) {
-      registerRoute('/privacy', () => {
-        setPageTitle('Privacy');
-        return app.getSequence('gotoPrivacySequence')();
-      });
+    registerRoute('/privacy', () => {
+      setPageTitle('Privacy');
+      return app.getSequence('gotoPrivacySequence')();
+    });
 
-      registerRoute('/contact', () => {
-        setPageTitle('Contact');
-        return app.getSequence('gotoContactSequence')();
-      });
-    }
+    registerRoute('/contact', () => {
+      setPageTitle('Contact');
+      return app.getSequence('gotoContactSequence')();
+    });
 
     registerRoute(
       '..',
