+++ web-client/src/routerPublic.js
@@ -1,4 +1,3 @@
-import { isCodeEnabled } from '../../codeToggles';
 import { setPageTitle } from './presenter/utilities/setPageTitle';
 import route from 'riot-route';
 
@@ -54,17 +53,15 @@ const router = {
       app.getSequence('gotoPublicSearchSequence')();
     });
 
-    if (isCodeEnabled(7142)) {
-      route('/privacy', () => {
-        setPageTitle('Privacy');
-        return app.getSequence('gotoPrivacySequence')();
-      });
+    route('/privacy', () => {
+      setPageTitle('Privacy');
+      return app.getSequence('gotoPrivacySequence')();
+    });
 
-      route('/contact', () => {
-        setPageTitle('Contact');
-        return app.getSequence('gotoContactSequence')();
-      });
-    }
+    route('/contact', () => {
+      setPageTitle('Contact');
+      return app.getSequence('gotoContactSequence')();
+    });
 
     route('..', () => {
       setPageTitle('Error');
