+++ web-client/src/views/AppComponent.jsx
@@ -80,7 +80,6 @@ import { UserContactEdit } from './UserContactEdit';
 import { UserContactEditProgress } from './UserContactEditProgress';
 import { WebSocketErrorModal } from './WebSocketErrorModal';
 import { WorkQueue } from './WorkQueue';
-import { applicationContext } from '../applicationContext';
 import { connect } from '@cerebral/react';
 import { state } from 'cerebral';
 import React, { useEffect } from 'react';
@@ -162,8 +161,6 @@ const pages = {
   WorkQueue,
 };
 
-const featureShowFooter = applicationContext.isCodeEnabled(7142);
-
 /**
  * Root application component
  */
@@ -190,10 +187,6 @@ export const AppComponent = connect(
 
     useEffect(() => {
       focusMain();
-      if (!featureShowFooter) {
-        const templateDateElement = document.getElementById('last-deployed');
-        templateDateElement.classList.remove('hide');
-      }
     }, [currentPage]);
 
     const CurrentPage = pages[currentPage];
@@ -216,7 +209,7 @@ export const AppComponent = connect(
         </main>
         <Loading />
 
-        {featureShowFooter && <Footer />}
+        <Footer />
 
         {showModal === 'TrialSessionPlanningModal' && (
           <TrialSessionPlanningModal />
