+++ web-client/src/views/AppComponentPublic.jsx
@@ -11,7 +11,6 @@ import { PublicPrintableDocketRecord } from './Public/PublicPrintableDocketRecor
 import { PublicSearch } from './Public/PublicSearch';
 import { TodaysOpinions } from './Public/TodaysOpinions';
 import { UsaBanner } from './UsaBanner';
-import { applicationContextPublic } from '../applicationContextPublic';
 import { connect } from '@cerebral/react';
 import { state } from 'cerebral';
 import React, { useEffect } from 'react';
@@ -27,8 +26,6 @@ const pages = {
   PublicSearch,
   TodaysOpinions,
 };
-
-const featureShowFooter = applicationContextPublic.isCodeEnabled(7142);
 
 /**
  * Root application component for the public site
@@ -47,10 +44,6 @@ export const AppComponentPublic = connect(
 
     useEffect(() => {
       focusMain();
-      if (!featureShowFooter) {
-        const templateDateElement = document.getElementById('last-deployed');
-        templateDateElement.classList.remove('hide');
-      }
     });
 
     const CurrentPage = pages[currentPage];
@@ -72,7 +65,7 @@ export const AppComponentPublic = connect(
         </main>
         <Loading />
 
-        {featureShowFooter && <Footer />}
+        <Footer />
       </React.Fragment>
     );
   },
